+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog
 =========
 
+unreleased
+-----------------------------------------
+
+* changed DEFAULT behavior in test config loader to correspond to Montague 0.1.6. This is a breaking change.
+
 0.1.0 (2014-11-12)
 -----------------------------------------